--- .eslintignore
@@ -1,3 +1,0 @@
-.next
-node_modules
-coverage
