+++ package.json
@@ -33,19 +33,15 @@
     "@emotion/styled": "^11.13.0",
     "@faker-js/faker": "^8.4.1",
     "@hookform/resolvers": "^3.3.4",
-    "@mui/base": "^5.0.0-beta.69",
-    "@mui/icons-material": "^5.16.7",
-    "@mui/lab": "^5.0.0-alpha.173",
-    "@mui/material": "^5.16.7",
+    "@mui/icons-material": "^7.3.5",
+    "@mui/lab": "7.0.0",
+    "@mui/material": "^7.3.5",
     "@mui/material-nextjs": "^7.3.6",
     "@mui/system": "^5.16.7",
-    "@mui/x-date-pickers": "^7.22.2",
+    "@mui/utils": "^7.3.5",
+    "@mui/x-date-pickers": "^8.18.0",
     "@next/bundle-analyzer": "^15.0.3",
     "@next/third-parties": "16.2.4",
-    "@storybook/addon-essentials": "^8.0.5",
-    "@storybook/addon-interactions": "^8.0.5",
-    "@storybook/addon-links": "^8.0.5",
-    "@storybook/addon-onboarding": "^8.0.5",
     "@tanstack/react-query": "^5.51.23",
     "@tanstack/react-table": "^8.20.6",
     "@types/react-syntax-highlighter": "^15.5.13",
@@ -60,11 +56,6 @@
     "dayjs": "^1.11.11",
     "deep-equal": "^2.2.3",
     "dotenv": "^17.2.3",
-    "eslint-config-next": "14.1.4",
-    "eslint-plugin-import": "^2.32.0",
-    "eslint-plugin-jsx-a11y": "^6.10.2",
-    "eslint-plugin-react": "^7.37.5",
-    "eslint-plugin-react-hooks": "^4.3.0",
     "framer-motion": "^12.18.1",
     "immer": "^10.1.1",
     "js-cookie": "^3.0.5",
@@ -72,11 +63,11 @@
     "lodash.get": "^4.4.2",
     "maildev": "^2.2.1",
     "motion": "^11.11.17",
-    "next": "^15.5.15",
-    "next-intl": "^3.10.0",
+    "next": "16.2.4",
+    "next-intl": "^4.9.1",
     "pretty-bytes": "^6.1.1",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "19.2.5",
+    "react-dom": "19.2.5",
     "react-hook-form": "^7.51.1",
     "react-markdown": "^9.1.0",
     "react-syntax-highlighter": "^16.1.0",
@@ -86,7 +77,6 @@
     "zustand": "^4.5.2"
   },
   "devDependencies": {
-    "@chromatic-com/storybook": "^3.2.2",
     "@next/bundle-analyzer": "16.2.4",
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/commit-analyzer": "^13.0.0",
@@ -95,10 +85,6 @@
     "@semantic-release/github": "11.0.1",
     "@semantic-release/npm": "^12.0.1",
     "@semantic-release/release-notes-generator": "^14.0.1",
-    "@storybook/blocks": "^8.0.5",
-    "@storybook/nextjs": "^8.0.5",
-    "@storybook/react": "^8.0.5",
-    "@storybook/test": "^8.0.5",
     "@testing-library/jest-dom": "^6.4.2",
     "@testing-library/react": "^16.0.0",
     "@testing-library/user-event": "^14.5.2",
@@ -112,21 +98,18 @@
     "@types/react": "19.2.14",
     "@types/react-dom": "19.2.3",
     "@types/testing-library__cypress": "^7.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.16.0",
-    "@typescript-eslint/parser": "^8.50.1",
+    "@typescript-eslint/eslint-plugin": "^8.59.0",
+    "@typescript-eslint/parser": "^8.59.0",
     "axe-core": "^4.10.3",
     "chromatic": "^11.18.0",
     "css-mediaquery": "^0.1.2",
     "cypress": "^15.8.1",
     "cypress-axe": "^1.7.0",
     "cypress-maildev": "^1.3.5",
-    "eslint": "^8",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-security": "^2.1.1",
-    "eslint-plugin-storybook": "^0.6.15",
-    "eslint-plugin-unused-imports": "^4.1.4",
+    "eslint": "^9.39.4",
+    "eslint-config-next": "^16.2.4",
+    "eslint-plugin-prettier": "^5.5.5",
+    "eslint-plugin-unused-imports": "^4.4.1",
     "express": "^4.22.1",
     "heroku-ssl-redirect": "^0.1.1",
     "jest": "^29.7.0",
@@ -136,7 +119,6 @@
     "next-router-mock": "^0.9.13",
     "prettier": "^3.2.5",
     "semantic-release": "^24.2.1",
-    "storybook": "^8.0.5",
     "ts-node": "^10.9.2",
     "typescript": "^5.9.3"
   },
