+++ src/app/api/auth/me/route.ts
