+++ src/components/DateInput/DateInput.test.tsx
@@ -1,5 +1,5 @@
 import { LocalizationProvider } from "@mui/x-date-pickers";
-import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFnsV3";
+import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFns";
 import "@testing-library/jest-dom";
 import { enGB } from "date-fns/locale/en-GB";
 import { render, screen } from "../../utils/testUtils";
