+++ src/components/DateInput/DateInput.tsx
@@ -3,8 +3,9 @@ import {
   LocalizationProvider,
   PickerChangeHandlerContext,
 } from "@mui/x-date-pickers";
-import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFnsV3";
-import { DatePicker, DatePickerProps } from "@mui/x-date-pickers/DatePicker";
+import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFns";
+import { DatePicker } from "@mui/x-date-pickers";
+import type { DatePickerProps } from "@mui/x-date-pickers";
 import { useLocale } from "next-intl";
 import { enGB } from "date-fns/locale/en-GB";
 import dayjs from "dayjs";
