+++ src/components/Mask/Mask.styles.ts
@@ -1,25 +1,20 @@
 import {
   AugmentedColorPaletteOptions,
   Box,
-  Theme,
   css,
   lighten,
   styled,
 } from "@mui/material";
 import { getAugmentedColor } from "../../utils/theme";
 
-export const StyledMask = styled(Box)(
-  ({
-    theme,
-    width,
-    height,
-    color = "primary",
-  }: {
-    theme: Theme;
-    width: string;
-    height: string;
-    color?: AugmentedColorPaletteOptions;
-  }) => css`
+interface MaskProps {
+  width: string;
+  height: string;
+  color?: AugmentedColorPaletteOptions;
+}
+
+export const StyledMask = styled(Box)<MaskProps>(
+  ({ theme, width, height, color = "primary" }) => css`
     background: ${getAugmentedColor(theme, color).main};
     padding: 2px;
     box-sizing: border-box;
