+++ src/utils/redirects.ts
@@ -3,10 +3,8 @@
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
 import { getRefreshAccessToken } from "@/app/actions/auth/getRefreshAccessToken";
-import { ROUTES } from "../consts/router";
-import { User } from "../types/application";
-import { getProfilePathByEntity } from "./entity";
 import { getLoginUrl, getRegisterUrl } from "./keycloak";
+import { getHomepageRedirectPath, isInPath } from "@/utils/redirect-utils";
 
 async function redirectToPath(redirectUrl: string, pathname?: string) {
   const cookieStore = await cookies();
@@ -20,16 +18,6 @@ async function redirectToPath(redirectUrl: string, pathname?: string) {
     redirect(redirectUrl);
   }
 }
-
-const getProfileRedirectPath = (user: User) => {
-  const redirectPath = getProfilePathByEntity(user);
-
-  return redirectPath;
-};
-
-const getRegisterRedirectPath = async () => {
-  return ROUTES.register.path;
-};
 
 async function getRefreshTokenRedirectPath() {
   const accessToken = await getRefreshAccessToken();
@@ -45,10 +33,6 @@ async function getRefreshTokenRedirectPath() {
   return null;
 }
 
-function getHomepageRedirectPath() {
-  return ROUTES.homepage.path;
-}
-
 async function redirectInvite() {
   return getRegisterUrl();
 }
@@ -57,28 +41,18 @@ async function getSeverErrorRedirectPath(
   accessToken: string | undefined,
   pathname: string
 ) {
+  const cookieStore = await cookies();
+  console.log(accessToken, "<----<<<");
   if (!accessToken) {
-    Cookies.set("redirectPath", pathname ?? "/", { path: "/" });
+    cookieStore.set("redirectPath", pathname ?? "/", { path: "/" });
   }
 
   return getHomepageRedirectPath();
 }
 
-function isInPath(pathname: string, pathToCompare: string | string[]) {
-  if (Array.isArray(pathToCompare)) {
-    return pathToCompare.some(item => item.includes(pathname));
-  }
-
-  return pathToCompare.includes(pathname);
-}
-
 export {
-  getHomepageRedirectPath,
-  getProfileRedirectPath,
   getRefreshTokenRedirectPath,
-  getRegisterRedirectPath,
   getSeverErrorRedirectPath,
-  isInPath,
   redirectInvite,
   redirectToPath,
 };
