+++ tsconfig.json
@@ -1,7 +1,11 @@
 {
   "compilerOptions": {
     "target": "ES2017",
-    "lib": ["dom", "dom.iterable", "esnext"],
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext"
+    ],
     "allowJs": true,
     "skipLibCheck": true,
     "strict": true,
@@ -9,10 +13,10 @@
     "noEmit": true,
     "esModuleInterop": true,
     "module": "esnext",
-    "moduleResolution": "node",
+    "moduleResolution": "bundler",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "jsx": "preserve",
+    "jsx": "react-jsx",
     "incremental": true,
     "plugins": [
       {
@@ -21,19 +25,33 @@
     ],
     "baseUrl": ".",
     "paths": {
-      "@/mocks/*": ["./mocks/*"],
-      "@/*": ["./src/*"]
+      "@/mocks/*": [
+        "./mocks/*"
+      ],
+      "@/*": [
+        "./src/*"
+      ]
     },
-    "types": ["cypress", "node", "jest", "@testing-library/jest-dom"]
+    "types": [
+      "cypress",
+      "node",
+      "jest",
+      "@testing-library/jest-dom"
+    ]
   },
-  "files": ["node_modules/gsap/types/index.d.ts"],
+  "files": [
+    "node_modules/gsap/types/index.d.ts"
+  ],
   "include": [
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
     ".storybook/*.tsx",
-    "cypress/e2e/users/affiliations/add.cy.ts"
+    "cypress/e2e/users/affiliations/add.cy.ts",
+    ".next/dev/types/**/*.ts"
   ],
-  "exclude": ["node_modules"]
+  "exclude": [
+    "node_modules"
+  ]
 }
